+++ dbentry/query.py
@@ -128,6 +128,16 @@ class MIZQuerySet(TextSearchQuerySetMixin, QuerySet):
             result[pk] = item_dict
         return result
 
+    def add_changelist_annotations(self):
+        """Add annotations expected by changelist views."""
+        if annotations := self.model.get_overview_annotations():
+            return self.annotate(**annotations)
+        return self
+
+    def overview(self) -> 'MIZQuerySet':
+        """Return a queryset that provides a comprehensive overview of the objects."""
+        return self.model.overview(self)
+
 
 class CNQuerySet(MIZQuerySet):
     # TODO: shouldn't get() update the name just like filter?