+++ tests/test_site/test_changelist.py
@@ -5,7 +5,7 @@ from django.test import override_settings
 from django.urls import reverse, path
 from django.views import View
 
-from dbentry.site.views import BaseListView
+from dbentry.site.views.base import BaseListView
 from tests.case import ViewTestCase, DataTestCase
 from tests.model_factory import make
 from tests.test_site.models import Band, Musician, Country
@@ -177,6 +177,7 @@ class TestBaseListView(ChangelistTestCase):
         The change page links in a row should contain the changelist request
         parameters.
         """
+        # NOTE: why? what's the benefit of this?
         request = self.get_request(data={'p': ['1']})
         view = self.get_view(request)
         obj = self.get_annotated_model_obj(self.obj)